transcriber/requirements.txt
groq
google-genai
pydub
requests